Add files via upload
Here I have asked the user to select either rock paper or scissors. And  also make the computer choose one out of three randomly. Now we will check who won the game and give them a score. This is done 5 times to say who has won finally and we will ask the user if they want to continue or reset and continue or exit the game.